feat: update API to use v1 endpoints and remove v2 references
